Modules/Options: Add support for array-based options
- This is an extension of the "conf" format which extends on top of the
  Lua config system. Now you can create array-options for your JS
  "options", so that you don't have to place long options on the same
  line in the conf file. You can now place values in arrays instead, and
  append their values to the array one by one, in a clear and easy way.

To define an array (in JS):
```
    var userConfig = new Options.advanced_options({
        // Values can be pre-defined here, or can be left empty. User
        // provided values will be appended to this array at runtime.
        // Note that ALL user-provided array values will be strings!
        presets: []
    });
```

To add values to the array (in .conf):
```
presets[]=This is an array entry
presets[]=This is another array entry
```

The trailing `[]` after each key is very important. Without it, the
value is invalid and will be ignored (with a warning in your console).

Result: `userConfig.getValue('presets')` will be an array containing:
`["This is an array entry","This is another array entry"]`.

There's also the `getMultiValue()` method if you want to define all
values as one long string within a single option (such as what's used
in the Blackbox multi-key bindings).

Lastly, there's the regular Lua-style way of reading all options at
once, which now supports arrays: `Options.read_options({ presets:[] })`.