Modules/SelectionMenu: Add support for object-items
- Previously, you had to define every menu option as a simple string. To
  attach special information which you wanted later, you either needed
  to store the option offset yourself and compare it to some other array
  where you kept your extra data, or you had to somehow embed the data
  in the string and then extract it later.

  This commit gives you a better solution. You can now use object-based
  option items. In that case, you simply have to define an object which
  MUST have the `menuText` key. All other object keys can be anything
  you want. But you should never create any keys with `menu*` in the
  name, since the menu will always use that prefix for any future
  extensions/special object keys.

Example of an object-based option:

```
var menu = new SelectionMenu();
menu.setCallbackMenuOpen(function(){ dump(menu.getSelectedItem()); });
menu.setOptions([
    'An old-school string-based option',
    {
        menuText: 'An object-based option',
        somethingExtra: [ 1, 2, 3 ]
    }
]);
menu.renderMenu();
```

The example dumps the selection to the terminal whenever you press
enter (the default "Open" key) on a menu selection. So be sure to run
mpv from your terminal, such as with `mpv --idle --force-window`.

It uses both string-based and object-based options for demonstration
purposes, but you are actually also allowed to do that in your real code
if you want to. The `getSelectedItem()` method will return the entire
selected value. In case of a string option, it returns the string. In
case of an object-based option it returns the whole object.

As you can see, this new feature makes it very easy to embed advanced
data within your options tree.